Shard input Z-sets in the context of workers.
Our initial implementation of `CollectionHandle` partitioned input
records across workers based on the hash of the key.  The Nexmark
benchmark taught us that doing this sequentially (in the client
thread) can become a performance bottleneck.  We therefore change
the `CollectionHandle` implementation to split records across workers
using round robin scheduling, which ensures even distribution, but
does not guarantee that all updates for the same key are sent to the
same worker; hence workers may need to re-partition the inputs,
which is more efficient than doing it in the client thread,
as work will be evenly split across workers.

Round robin does not work for upserts, which must be processed in
the same order they were submitted by the client.  We therefore
introduce a new type of input handle, `UpsertHandle`, which
consistently maps keys to workers.