Support for "shardedness-preserving" operators.
At the moment operators that require sharded inputs will blindly shard
their input streams.  This is often unnecessary as the stream may
already be sharded because one of the following conditions holds:

- The upstream operator always produces sharded outputs
  (e.g., `distinct`, `aggregate`)

- The upstream operator is shardedness-preserving, i.e., produces
  sharded outputs given sharded inputs (e.g., `filter`), and its
  input stream happens to be sharded.

- The input is sharded for some other reason known to the user,
  e.g., the user pushes sharded batches to the `InputHandle`, so
  the input stream it produces is sharded.

We add the infrastructure to track this information through the
dataflow graph and to skip the sharding step if the inputs is
known to be sharded.