[nexmark] Optimized implementation of q15.
q15 computes several different aggregates over the input stream.  The
initial implementation computes them all simultaneously, which requires
scanning the entire input relation on each input (since the relation is
grouped by date, which is the same in all input records), which is
completely non-incremental.

The new implementation is fully incremental by computing each aggregate
separately and decomposing it into a sequence of incremental
computations:

1. filter input stream (for aggregates that work on a subset of inputs)
2. project on relavant fields
3. `distinct_incremental` (for `DISTINCT` aggregates)
4. `aggregate_linear` - luckily, all aggregates in this query are `count`,
   which are linear.

The results must be joined in a single output stream, which required
implementing an outer join.