[nexmark] Fixup q8.
We could extract window bounds from the `window_bounds` stream by
converting it into a batch with a single value and joining this batch
with the output stream.  But for a tumbling window the solution is even
simpler, since window bounds can be computed from individual data points
by rounding their timestamps down to `TUMBLE_SECONDS`.