Rolling time series aggregation.
Rolling aggregation computes an aggregate over a relative time range
(e.g., the last three months) for each event in the input stream.
I had to build a lot of infrastructure to implement it efficiently (no
benchmarks yet, so not sure how efficient it is after all).  The
brute-force approach of computing the entire aggregate from scratch for
each event can be extremely expensive as it requires iterating  over a
large volume of data for each event.  Out-of-order inputs make things
worse, as they require re-computing the aggregate multiple times for
each record.  Much of this work is redundant, as the contents of the
3-months time window for adjacent events is nearly identical, so it
should be possible to reuse most of the computation.  We achieve this by
covering the entire timeline with a radix tree, where each node stores
the value of the aggregate over the time range determined by its bit
prefix.  Given this tree, we can compute the aggregate over arbitrary
time range of length `n` by visiting `O(log(n))` nodes.

So we separate radix tree computation into a separate operator under
`time_series/radix_tree` and use it as building blocks for the rolling
aggregation operator.