Remove `TraceReader` trait.
`TraceReader` only defined a single method, `map_batches`.  Work on
persistent traces showed that this is not a good design, as it exposes
the internal structure of the trace.  We only used this API in one place
in `distinct.rs`, and that code would have to be re-written anyway as we
generalize `distinct` to work in arbitrary nested scopes.

We eliminate the `map_batches` method from the public API and downgrade
it to a private method of `Spine`.  We also remove the no longer useful
trait `TraceReader`.