Simplified endpoint configuration.
We simplify the input endpoint configuration format, which tried to be
overly general, introducing unnecesary complexity and potential user
confusion.

Background: a typical input endpoint (e.g., Kafka stream or file)
connects to a single input stream of the circuit.  In the future we will
introduce new kinds of transport such as database CDC streams, which
can push data to mulitple input streams of the circuit.  In order to
support these future use cases, we did not standardize the
endpoint-to-stream binding syntax, leaving it to parser-specific
configuration.  This resulted in clumsy internal API for adapters,
confusing external API and in the future potentially confusing UI
experience.

In this commit we simplify the model, moving binding specification out
of the parser config and into the endpoint config, which now has the
following structure (`stream` is the name of the input stream of the
circuit that the endpoint connects to):

```
stream: test_input
transport:
    name: file
    config:
        ...
format:
    name: csv
    config:
        ...
```

It shouldn't be hard to generalize it in the future to support adapters
with multiple input streams.