Get default broker address from env variable.
We would like to avoid hardwiring Kafka broker address in pipeline
configs.  The concrete use case is running the same demo scripts
with and without docker.

We modify the Kafka transport to use the value of the `REDPANDA_BROKERS`
env variable when the endpoint config doesn't explicitly specify the
`bootstrap.servers` parameter value.