Fixup Redpanda container config.
I copied the docker compose script from redpanda documentation to get it
to work.  I don't fully understand each option yet, but the crux of it
is that it's not sufficient to expose the redpanda API port outside of
the continer.  The Redpanda broker also needs to know its addresses both
on the docker network and on the extrnal network.