Move timekeeping to the circuit.
We get rid of another piece of technical debt, related to the fact that
circuits did not do their own time keeping.  As a result, any operator that
needed to know the current time had to run its own clock.  More
importantly, any operator that recorded input batches in a trace
had to pick its own timestamp type for the trace.  This choice
bubbled up all the way to the user, so that incremental stateful
operators ended up with an extra generic parameter that had to be
specified manually, making the API more complex and error-prone (since
nothing prevented the user from picking the wrong timestamp type).

The solution implenented in this commit:

- `trait WithClock` - represents an object with a clock whose type is
  given by the associated type `Time`.

- Abstract away the circuit API into a trait `trait Circuit: WithClock`.
  We still only have one implementation of this trait, but this allows
  us to model `Time` and `Parent` as associated types of the trait.

- Derive the timestamp type of all stateful operators from the circuit,
  which eliminates the extra type argument to `join`, `aggregate`,
  `distinct`, etc.  All these operators can now be invoked without
  explicitly type arguments in most cases.

- We setup a fixed hierarchy of timestamp types, induced by the
  `WithClock::Nested` associate type:
  - `()` - root circuit
  - `NestedTimestamp32` - first-level nested circuit
  - `Product<>` - >=2-level nested circuit.
  If this proves too restrictive, we can generalize this architecture
  by allowing the user to explicitly pick a timestamp type when
  instantiating a (nested) circuit.