Bounded-memory `window` operator.
We build the infrastructure to garbage collect unused prefixes of
Z-set traces.

Motivation
==========

Window-based operators (e.g., `window`, rolling aggregate) maintain the trace
of their input and/or output collections, but only use the suffix of the trace
within the current window (more precisely, they need to look as far back
as the width of the window plus the largest possible out-of-ordedness of
the time series stream, aka "lateness").  In order to keep the memory of
such operators bounded we need to truncate the prefix of the trace
outside the window.  Note that this is semantically different from
removing unused keys: the keys are still logically in the trace, but
we know that they will never be accessed and hence don't need to store
them.

Design
======

We rely on the `BatchReader::truncate_keys_below` API introduced in an
earlier commit to garbage collect unused keys.  One complication arises
from the fact that a trace can be used by multiple downstream operators,
each with its own retention threshold.  E.g., if a collection is
aggregated using two different time windows, we need to keep sufficient
state to maintain the wider window.  Moreover, if the collection is used
by an operator like `join` that requires unbounded trace, then
truncation should be disabled altogether.  To support these use cases,
we allow each downstream consumer to specify their own lower bound and
use the minimum of all bounds as the effective bound of the trace

API
===

- We implement the truncation logic in the `Z1Trace` operator, which
  computes and enforces the effective bound at the end of each clock
  cycle.

- We provide new flavors of `trace` and `integrate_trace` operators that
  take the user-provided bound and pass it on to `Z1Trace`.

- We change the `window` operator to use the new feature to
  truncate its input trace and this bound its memory usage.

- The new `partitioned_rolling_aggregate_with_watermark` operator takes
  advantage of limited out-of-ordedness of its input stream to bound
  its memory usage.  The initial implementation bounds the input trace
  and the size of the radix tree by applying `window` to the input
  stream.  It doesn't bound the output trace yet.  This is coming in
  the next commit.

Along the way we also change the signature of the `window` operator.
The original implementation took an indexed Z-set indexed by time and
output a Z-set containing values only.  The new API keeps the timestamp
from the input stream, which seems to reflect the more common usage of
the API.